Check In + mensaje de error al login

+++ src/com/company/Models/Payment.java
@@ -2,13 +2,13 @@ package com.company.Models;
 
 public class Payment {
 
-    private int days;
+    private long days;
     private float price;
     private Guest guest;
 
     public Payment (){}
 
-    public Payment (int days, float price, Guest guest){
+    public Payment (long days, float price, Guest guest){
         this.days = days;
         this.price = price;
         this.guest = guest;
